Fix idempotence of attributes
we're not allowed to prepend hardline in general, since topiary doesn't perform it before checking if the surrounding expr is multiline: it makes it non-idempotent on single-line exprs that contain an attribute.

The other solution would be to match attrs explicitely in the appropriate contexts to turn all softlines to hardlines explicitely in that case. Let's not bother.